backend/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# AI
google-generativeai==0.3.2

# File handling
python-multipart==0.0.6

# Image processing and storage
Pillow==10.1.0
cloudinary==1.37.0
